Confidence Level Created
Still building the level.

To do:
Make room for Confidence.
Make confidence puzzle(s).

Added several rooms for the player to go in, to move forward they must do little tasks in the room.

Also solution to the button puzzle is 1 2 4.  Maybe we can put a hint somewhere to the player for it?